+++ packages/vue-dot/src/elements/ExternalLinks/ExternalLinks.vue
@@ -1,5 +1,6 @@
 <template>
 	<VMenu
+		ref="menu"
 		v-model="menu"
 		v-bind="options.menu"
 		:top="bottom"
@@ -82,7 +83,7 @@
 	import { PositionEnum } from './PositionEnum';
 
 	import { ExternalLink, Position } from './types';
-	import { IndexedObject } from '../../types';
+	import { IndexedObject, Refs } from '../../types';
 
 	import { customizable } from '../../mixins/customizable';
 
@@ -138,6 +139,11 @@
 			fixed: {
 				type: Boolean,
 				default: false
+			},
+			/** Apply to set position's VMenu with 12px: default value of Vuetify */
+			isDefault: {
+				type: Boolean,
+				default: false
 			}
 		}
 	});
@@ -146,6 +152,13 @@
 
 	@Component
 	export default class ExternalLinks extends MixinsDeclaration {
+		// Extend $refs
+		$refs!: Refs<{
+			menu: {
+				calculatedLeft: string
+			};
+		}>;
+
 		locales = locales;
 
 		linkIcon = mdiOpenInNew;
@@ -181,8 +194,11 @@
 
 		get menuClass(): string {
 			const positionClass = this.right ? 'right-0' : 'left-0';
-
-			return `vd-external-links-menu ${positionClass}`;
+            let contentClass = 'vd-external-links-menu';
+            if(this.isDefault){
+                contentClass = `${contentClass} ${positionClass}`;
+            }
+			return contentClass;
 		}
 
 		get btnTextSpacing(): string {