+++ packages/vue-dot/src/elements/FranceConnectBtn/index.ts
@@ -1,0 +1,3 @@
+import FranceConnectBtn from './FranceConnectBtn.vue';
+
+export default FranceConnectBtn;