+++ packages/vue-dot/src/elements/FranceConnectBtn/locales.ts
@@ -1,0 +1,4 @@
+export const locales = {
+	btnText: 'S’identifier avec France Connect',
+	href: 'https://app.franceconnect.gouv.fr/'
+};